Split off GLesmos
