+++ src/main/metadata/interface.ts
@@ -1,6 +1,6 @@
 export default interface Metadata {
   version: 2;
-  expressions: Record<string, Expression>;
+  expressions: Record<string, Expression | undefined>;
 }
 
 export interface Expression {
