+++ src/preload/moduleOverrides/glesmos.replacements
@@ -52,7 +52,7 @@ __rest__,
 $DCGView.createElement(
   $DCGView.Components.If,
   {
-    predicate: () => DesModder.controller.canBeGLesmos(this.id),
+    predicate: () => DSM.glesmos?.canBeGLesmos(this.id),
   },
   () => $DCGView.createElement(
     "div",
@@ -60,8 +60,8 @@ $DCGView.createElement(
     () => DesModder.controller.format("GLesmos-label-toggle-glesmos"),
     $DCGView.createElement($ToggleView, {
       ariaLabel: () => DesModder.controller.format("GLesmos-label-toggle-glesmos"),
-      toggled: () => window.DesModder?.controller?.isGlesmosMode?.(this.id),
-      onChange: () => window.DesModder?.controller?.toggleGlesmos?.(this.id),
+      toggled: () => DSM.glesmos?.isGlesmosMode?.(this.id),
+      onChange: () => DSM.glesmos?.toggleGlesmos?.(this.id),
     })
   )
 )
@@ -103,8 +103,8 @@ __rest__,
 $DCGView.createElement(
   $DCGView.Components.If,
   {
-    predicate: () => DesModder.controller.canBeGLesmos(this.id) &&
-      !DesModder.controller.isInequality(this.id)
+    predicate: () => DSM.glesmos?.canBeGLesmos(this.id) &&
+      !DSM.glesmos?.isInequality(this.id)
   },
   () => $DCGView.createElement(
     "div",
@@ -112,8 +112,8 @@ $DCGView.createElement(
     () => DesModder.controller.format("GLesmos-label-toggle-glesmos"),
     $DCGView.createElement($ToggleView, {
       ariaLabel: () => DesModder.controller.format("GLesmos-label-toggle-glesmos"),
-      toggled: () => window.DesModder?.controller?.isGlesmosMode?.(this.id),
-      onChange: () => window.DesModder?.controller?.toggleGlesmos?.(this.id),
+      toggled: () => DSM.glesmos?.isGlesmosMode?.(this.id),
+      onChange: () => DSM.glesmos?.toggleGlesmos?.(this.id),
     })
   )
 )
@@ -149,8 +149,8 @@ $DCGView.createElement(
   $DCGView.Components.If,
   {
     predicate: () =>
-      DesModder.controller.isGlesmosMode(this.id) &&
-      DesModder.controller.isInequality(this.id),
+      DSM.glesmos?.isGlesmosMode(this.id) &&
+      DSM.glesmos?.isInequality(this.id),
   },
   () => $DCGView.createElement(
     "div",
@@ -158,13 +158,13 @@ $DCGView.createElement(
     () => DesModder.controller.format("GLesmos-confirm-lines"),
     $DCGView.createElement($ToggleView, {
       ariaLabel: () => DesModder.controller.format("GLesmos-confirm-lines"),
-      toggled: () => window.DesModder?.controller?.isGLesmosLinesConfirmed?.(this.id),
-      onChange: () => window.DesModder?.controller?.toggleGLesmosLinesConfirmed?.(this.id),
+      toggled: () => DSM.glesmos?.isGLesmosLinesConfirmed?.(this.id),
+      onChange: () => DSM.glesmos?.toggleGLesmosLinesConfirmed?.(this.id),
     }),
     $DCGView.createElement(
       $DCGView.Components.If,
       {
-        predicate: () => !DesModder?.controller?.isGLesmosLinesConfirmed?.(this.id)
+        predicate: () => !DSM.glesmos?.isGLesmosLinesConfirmed?.(this.id)
       },
       () => $DCGView.createElement("div",
         { class: $DCGView.const("dsm-gl-lines-confirm-body") },
@@ -221,8 +221,8 @@ addStatement($stmt) {
 *Replace* `body` with
 ```js
 if ($stmt.type === "statement") {
-  $stmt.glesmos = window.DesModder?.controller?.isGlesmosMode($stmt.id);
-  $stmt.glesmosLinesConfirmed = window.DesModder?.controller?.isGLesmosLinesConfirmed($stmt.id);
+  $stmt.glesmos = DSM.glesmos?.isGlesmosMode($stmt.id);
+  $stmt.glesmosLinesConfirmed = DSM.glesmos?.isGLesmosLinesConfirmed($stmt.id);
 }
 __body__
 ```
